Move tests from test_jobs to test_job
